Move(Move Exceptions): 공통 사용 exception들 경로 이동
