+++ src/main/java/com/sparta/newsfeed/common/exception/InvalidUserException.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.user;
+package com.sparta.newsfeed.common.exception;
 
 public class InvalidUserException extends RuntimeException {
     private static final String MESSAGE = "유효한 유저가 아닙니다.";
