+++ src/main/java/com/sparta/newsfeed/common/exception/NotFoundUserException.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.user;
+package com.sparta.newsfeed.common.exception;
 
 public class NotFoundUserException extends RuntimeException {
     private static final String MESSAGE = "유저를 찾을 수 없습니다.";
