+++ src/main/java/com/sparta/newsfeed/profile/ProfileService.java
@@ -2,8 +2,8 @@ package com.sparta.newsfeed.profile;
 
 import com.sparta.newsfeed.profile.dto.ProfileRequestDto;
 import com.sparta.newsfeed.profile.dto.ProfileResponseDto;
-import com.sparta.newsfeed.user.InvalidUserException;
-import com.sparta.newsfeed.user.NotFoundUserException;
+import com.sparta.newsfeed.common.exception.InvalidUserException;
+import com.sparta.newsfeed.common.exception.NotFoundUserException;
 import com.sparta.newsfeed.user.User;
 import com.sparta.newsfeed.user.UserRepository;
 import lombok.RequiredArgsConstructor;
